requirements.txt
opencv-python
numpy
pillow
torch
torchvision
clip-anytorch
faiss-cpu
pandas
transformers
roboflow
